separate dev error tests
