+++ test/e2e/app-dir/metadata-dynamic-routes/index.test.ts
@@ -444,86 +444,6 @@ describe('app dir - metadata dynamic routes', () => {
   })
 
   if (isNextDev) {
-    it('should error when id is missing in generateImageMetadata', async () => {
-      const iconFilePath = 'app/metadata-base/unset/icon.tsx'
-      const contentMissingIdProperty = `
-      import { ImageResponse } from 'next/og'
-      export async function generateImageMetadata() {
-        return [
-          {
-            contentType: 'image/png',
-            size: { width: 48, height: 48 },
-            // id: 100,
-          },
-          {
-            contentType: 'image/png',
-            size: { width: 48, height: 48 },
-            id: 101,
-          },
-        ]
-      }
-
-      export default function icon() {
-        return new ImageResponse(<div>icon</div>)
-      }
-      `
-
-      const originalOutputIndex = next.cliOutput.length
-
-      await next.patchFile(iconFilePath, contentMissingIdProperty)
-      await next.fetch('/metadata-base/unset/icon/100')
-
-      try {
-        await check(async () => {
-          expect(next.cliOutput.substring(originalOutputIndex)).toContain(
-            `id property is required for every item returned from generateImageMetadata`
-          )
-          return 'success'
-        }, /success/)
-      } finally {
-        await next.deleteFile(iconFilePath)
-        await next.fetch('/metadata-base/unset/icon/100')
-      }
-    })
-
-    it('should error when id is missing in generateSitemaps', async () => {
-      const sitemapFilePath = 'app/metadata-base/unset/sitemap.tsx'
-      const contentMissingIdProperty = `
-      import { MetadataRoute } from 'next'
-
-      export async function generateSitemaps() {
-        return [
-          { },
-        ]
-      }
-
-      export default function sitemap({ id }): MetadataRoute.Sitemap {
-        return [
-          {
-            url: 'https://example.com/',
-            lastModified: '2021-01-01',
-          },
-        ]
-      }`
-
-      const originalOutputIndex = next.cliOutput.length
-
-      await next.patchFile(sitemapFilePath, contentMissingIdProperty)
-      await next.fetch('/metadata-base/unset/sitemap/0.xml')
-
-      try {
-        await check(async () => {
-          expect(next.cliOutput.substring(originalOutputIndex)).toContain(
-            `id property is required for every item returned from generateSitemaps`
-          )
-          return 'success'
-        }, /success/)
-      } finally {
-        await next.deleteFile(sitemapFilePath)
-        await next.fetch('/metadata-base/unset/sitemap/0.xml')
-      }
-    })
-
     it('should error if the default export of dynamic image is missing', async () => {
       const ogImageFilePath = 'app/opengraph-image.tsx'
       const ogImageFileContent = await next.readFile(ogImageFilePath)
